{"schema":"decodex/commit/1","summary":"Split MCP server core tool router into tools module after strict validation.","authority":"manual"}
