+++ apps/elf-mcp/src/app/server/tools.rs
@@ -1,2 +1,3 @@
 mod admin;
+mod core;
 mod docs;
